ui: add button to install requirements
