+++ obs_osc/__init__.py
@@ -1,0 +1,1 @@
+from .osc.server import OscServer
